requirements.txt
tensorflow>=2.10
pandas>=1.5
numpy>=1.23
matplotlib>=3.5
flask>=2.2
werkzeug>=2.2
Pillow>=9.0
